+++ packages/polysemy-log/lib/Polysemy/Log/Log.hs
@@ -1,7 +1,6 @@
 -- |Description: Internal
 module Polysemy.Log.Log where
 
-import qualified Data.Text.IO as Text
 import Polysemy (interpretH, runT)
 import Polysemy.Async (Async)
 import Polysemy.Conc (Race)
@@ -11,12 +10,11 @@ import Polysemy.Resource (Resource)
 import Polysemy.Time (GhcTime, interpretTimeGhc)
 
 import Polysemy.Log.Conc (interceptDataLogConc)
-import Polysemy.Log.Data.DataLog (DataLog(DataLog, Local), dataLog)
-import Polysemy.Log.Data.Log (Log(Log))
+import Polysemy.Log.Data.DataLog (DataLog (DataLog, Local), dataLog)
+import Polysemy.Log.Data.Log (Log (Log))
 import Polysemy.Log.Data.LogEntry (LogEntry, annotate)
 import Polysemy.Log.Data.LogMessage (LogMessage)
-import Polysemy.Log.Data.LogMetadata (LogMetadata(Annotated), annotated)
-import Polysemy.Log.Format (formatLogEntry)
+import Polysemy.Log.Data.LogMetadata (LogMetadata (Annotated), annotated)
 
 -- |Interpret 'Log' into the intermediate internal effect 'LogMetadata'.
 interpretLogLogMetadata ::
@@ -80,7 +78,7 @@ interpretLogDataLogConc maxQueued =
   raiseUnder2
 {-# INLINE interpretLogDataLogConc #-}
 
--- |Helper for maintaining context function as state that is applied to each logged message, allowing the context of a
+-- |Helper for maintaining a context function as state that is applied to each logged message, allowing the context of a
 -- block to be modified.
 interpretDataLogLocal ::
   ∀ a r .
@@ -102,54 +100,3 @@ interpretDataLog ::
   InterpreterFor (DataLog a) r
 interpretDataLog =
   interpretDataLogLocal id
-
--- |Interpret 'DataLog' by printing to stderr, converting messages to 'Text' with the supplied function.
-interpretDataLogStderrWith ::
-  Member (Embed IO) r =>
-  (a -> Text) ->
-  InterpreterFor (DataLog a) r
-interpretDataLogStderrWith fmt =
-  interpretDataLog \ msg -> embed (Text.hPutStrLn stderr (fmt msg))
-{-# INLINE interpretDataLogStderrWith #-}
-
--- |Interpret 'DataLog' by printing to stderr, converting messages to 'Text' by using 'Show'.
-interpretDataLogStderr ::
-  Show a =>
-  Member (Embed IO) r =>
-  InterpreterFor (DataLog a) r
-interpretDataLogStderr =
-  interpretDataLogStderrWith show
-{-# INLINE interpretDataLogStderr #-}
-
--- |Interpret 'Log' by printing to stderr, converting messages to 'Text' with the supplied function.
-interpretLogStderrWith ::
-  Members [Embed IO, GhcTime] r =>
-  (LogEntry LogMessage -> Text) ->
-  InterpreterFor Log r
-interpretLogStderrWith fmt =
-  interpretDataLogStderrWith fmt .
-  interpretTimeGhc .
-  interpretLogDataLog .
-  raiseUnder2
-{-# INLINE interpretLogStderrWith #-}
-
--- |Interpret 'Log' by printing to stderr, using the default formatter.
---
--- Since this adds a timestamp, it has a dependency on 'GhcTime'.
--- Use 'interpretLogStderr'' for a variant that interprets 'GhcTime' in-place.
-interpretLogStderr ::
-  Members [Embed IO, GhcTime] r =>
-  InterpreterFor Log r
-interpretLogStderr =
-  interpretLogStderrWith formatLogEntry
-{-# INLINE interpretLogStderr #-}
-
--- |Interpret 'Log' by printing to stderr, using the default formatter, then interpreting 'GhcTime'.
-interpretLogStderr' ::
-  Member (Embed IO) r =>
-  InterpreterFor Log r
-interpretLogStderr' =
-  interpretTimeGhc .
-  interpretLogStderr .
-  raiseUnder
-{-# INLINE interpretLogStderr' #-}